requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
PyYAML==6.0.1
aiomysql==0.2.0 